revert r6957 (refs #1180)
git-svn-id: http://svn.freetz.org/trunk@6959 f5190166-0702-4917-9039-51ec32eddaf5

+++ make/mod/files/root/usr/lib/mww/do_external_handler.sh
@@ -67,8 +67,12 @@ kill() {
 	fi
 	if $delete; then
 		echo -n "Removing old stuff ... "
-		rm -rf "$EXTERNAL_TARGET"
-		[ $? -ne 0 ] && echo "failed." || echo "done."
+		if [ ! -e $EXTERNAL_TARGET/.external ]; then
+			echo "$EXTERNAL_TARGET is not an external dir."
+		else
+			rm -rf "$EXTERNAL_TARGET"
+			[ $? -ne 0 ] && echo "failed." || echo "done."
+		fi
 	else
 		echo "Not deleting old external stuff."
 	fi
